Fix telescope config by using defaults map
This was a really annoying bug. The function wrapping I was doing to
pass in the same opts map to every builtins call was making those
funciton calls use the first buffer loaded as the only buffer operated
on because of some weird Lua scoping. The much cleaner way to pass in
the same args to every builtin is to use the defaults map.