Improve font installation during bootstrapping
I've decided to add the JetBrains Mono Nerd Fonts that I love to this
repo. I also made a bash script that installs them to the proper
location. Also kitty now looks for both the normal and nerd font
versions, and I fixed the dnf install command for the normal fonts.