Improve i3status-rs bar command invokation
This patch adds a script to start i3status-rs in Sway. The PATH of Sway
is very limited by default on Fedora (doesn't source the /etc/bashrc
file that Nix appends to) so I just made a script that adds to the PATH
so i3status-rs from Nix will be used if installed.